// src/app/precisa/analisis/Egreso/model/dtoegresodetalle.ts
export class DtoEgresoDetalle {
    
    constructor(){}
    num?:                 number;
    TipoDocumento?:       string;
    Documento?:           string;
    Proveedor?:           string;
    DesEstado?:           string;
    Id?:                  number;
    IdEgreso?:            number;
    IdProveedor?:         number;
    FechaRegistro?:       Date;
    DesTipoDocumento?:      string;
    Serie?:               string;
    NroDocumento?:        string;
    Concepto?:            string;
    TipoImpuesto?:        number;
    Cantidad?:            number;
    MontoAfecto?:         number;
    MontoImpuesto?:       number;
    MontoTotal?:          number;
    Archivo?:             string;
    Estado?:              number;
    UsuarioCreacion?:     string;
    UsuarioModificacion?: string;
    FechaCreacion?:       Date;
    FechaModificacion?:   Date;
    IpCreacion?:          string;
    IpModificacion?:      string;
    DesEstado1?:          string;
}